Something working
I currently have it working to a slight extent. You can docker compose up --build and see all three containers running. Problematically, once you "submit" a new shark, it crashes the app, but this is progress.

I was encountering tons of trouble locally; I don't really want to talk about it. Basically, the fix was to use the latest version of the node alpine image by docker. The Dockerfile from the tutorial repo needed to be modified (removed the USER node line) because I couldn't npm install mongoose with the newer alpine version.