Updating the JS in shark.js
Since I had to update the project (npm, mongoose, and ejs) to the latest versions, the tutorial code used unsupported methods; I have replaced them.

I also added our names to the "Shark Info" line. The app still crashes when you submit a shark, but this is a massive improvement.