Basic three-component setup, compose file
I have a basic set up, but it is not communicating with mongodb currently. I'll need to make that a dependency.

I have a front end being served on nginx, an api using node, and a db using mongo; currently the front end is being served, but I haven't got much else working.